% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/available.R
\name{available}
\alias{available}
\title{Available Packages and Functions}
\usage{
available()
}
\value{
Returns a named list of function vectors named by package.
}
\description{
See \pkg{rnltk}'s available packages and functions.
}
\examples{
\dontrun{available()}
}

